feat(side-sheet): remove emotion
