+++ packages/side-sheet/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@equinor/fusion-react-side-sheet",
-  "version": "0.1.0",
+  "version": "1.0.0-beta",
   "description": "React component for displaying a side-sheet",
   "keywords": [
     "side-sheet",
