+++ packages/side-sheet/src/components/SideSheet.tsx
@@ -1,107 +1,138 @@
-import { Scrim } from '@equinor/eds-core-react';
-import { Resizable } from 're-resizable';
-import { PropsWithChildren, useState } from 'react';
+import React, { useCallback } from 'react';
+
+import { Button, Icon } from '@equinor/eds-core-react';
+import { close as closeIcon } from '@equinor/eds-icons';
+import { tokens } from '@equinor/eds-tokens';
+import { type PropsWithChildren, useRef } from 'react';
+import { Actions } from './Actions';
+import { Content } from './Content';
+import { FullscreenIcon } from './icon/FullscreenIcon';
+import { HEXString, Indicator } from './Indicator';
+import { SubTitle } from './SubTitle';
+import { Title } from './Title';
+
+import { SideSheetBase, type SideSheetProps } from './SideSheetBase';
 import styled from 'styled-components';
 
-export const StyledScrim = styled(Scrim)`
-  animation: ScrimAnimation ease 0.3s;
-  animation-iteration-count: 1;
-  animation-fill-mode: forwards;
-
-  overflow: hidden !important;
-  @keyframes ScrimAnimation {
-    0% {
-      opacity: 0;
-    }
-    100% {
-      opacity: 1;
-    }
-  }
+const StyledFlexBox = styled.div`
+  display: flex;
 `;
 
-export const StyledSideSheet = styled.div`
-  height: 100vh;
-  position: fixed;
-  top: 0;
-  transition: right 10s;
-  animation: Animation ease 0.3s;
-  right: 0px;
-
-  @keyframes Animation {
-    0% {
-      right: -500px;
-    }
-    100% {
-      right: 0px;
-    }
-  }
+const StyledFlexColumn = styled(StyledFlexBox)`
+  flex-direction: column;
 `;
 
-export const StyledSideSheetContent = styled.div`
-  height: 100%;
-  background: #fff; // TODO: token
-  width: 100%;
+const StyledContainerWrapper = styled(StyledFlexColumn)`
+  background: ${tokens.colors.ui.background__default.hex};
 `;
 
-const MIN_WIDTH = 500;
-export type SideSheetProps = {
-  isOpen: boolean;
-  isDismissable?: boolean;
+const StyledHeader = styled(StyledFlexBox)`
+  padding: 1rem;
+  height: 48px;
+  justify-content: space-between;
+`;
+
+type PortalSideSheet = SideSheetProps & {
   minWidth?: number;
-  onClose(): void;
+  enableFullscreen?: boolean;
 };
 
-export const SideSheet = (props: PropsWithChildren<SideSheetProps>) => {
-  const { isOpen, onClose, isDismissable, minWidth, children } = props;
-  const [width, setWidth] = useState(minWidth ?? MIN_WIDTH);
+type SideSheetComponents = {
+  indicator?: React.ReactElement<unknown, string | React.JSXElementConstructor<{ color: HEXString }>>;
+  title?: React.ReactElement<unknown, string | React.JSXElementConstructor<{ title: string }>>;
+  subTitle?: React.ReactElement<unknown, string | React.JSXElementConstructor<{ subTitle: string }>>;
+  actions?: React.ReactElement<unknown, string | React.JSXElementConstructor<PropsWithChildren<unknown>>>;
+  content?: React.ReactElement<unknown, string | React.JSXElementConstructor<PropsWithChildren<unknown>>>;
+};
+
+Icon.add({
+  close: closeIcon,
+});
+
+export const SideSheet = (props: PropsWithChildren<PortalSideSheet>) => {
+  const { onClose, children, enableFullscreen } = props;
+
+  const ref = useRef<HTMLDivElement>(null);
+  const handleFullscreenClick = useCallback(() => {
+    if (document.fullscreenElement) {
+      document.exitFullscreen().catch((err) => console.error(err));
+    } else {
+      ref.current?.requestFullscreen();
+    }
+  }, [ref]);
+
+  const components: SideSheetComponents = {
+    indicator: undefined,
+    title: undefined,
+    subTitle: undefined,
+    actions: undefined,
+    content: undefined,
+  };
+
+  React.Children.forEach(children, (child) => {
+    if (!React.isValidElement(child)) return;
+    if (child.type === Indicator) {
+      components.indicator = child;
+    } else if (child.type === Title) {
+      components.title = child;
+    } else if (child.type === SubTitle) {
+      components.subTitle = child;
+    } else if (child.type === Content) {
+      components.content = child;
+    } else if (child.type === Actions) {
+      components.actions = React.cloneElement(child as React.ReactElement<any>, {
+        sideSheetRef: ref,
+      });
+    } else {
+      throw Error(`unsupported child ${child.type} in SideSheet component`);
+    }
+  });
+
+  if (!components.title) {
+    throw Error('Title Component is required child');
+  }
+
+  if (!components.subTitle) {
+    throw Error('SubTitle Component is required child');
+  }
+
+  if (!components.content) {
+    throw Error('Content Component is required child');
+  }
 
   return (
-    <StyledScrim open={isOpen} onClose={onClose} isDismissable={isDismissable}>
-      <StyledSideSheet>
-        <Resizable
-          size={{ width, height: '100%' }}
-          maxWidth={'100vw'}
-          minWidth={minWidth ?? MIN_WIDTH}
-          onResizeStop={(e, direction, ref, d) => {
-            e.stopPropagation();
-            e.stopImmediatePropagation();
-            setWidth(width + d.width);
-          }}
-          handleComponent={{
-            left: (
-              <svg width="14" height="25" viewBox="0 0 14 25" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <rect
-                  x="7.21924"
-                  y="21.274"
-                  width="17"
-                  height="2"
-                  transform="rotate(-90 7.21924 21.274)"
-                  fill="#6F6F6F"
-                />
-                <rect
-                  x="11.2192"
-                  y="21.274"
-                  width="17"
-                  height="2"
-                  transform="rotate(-90 11.2192 21.274)"
-                  fill="#6F6F6F"
-                />
-              </svg>
-            ),
-          }}
-          handleStyles={{
-            left: {
-              position: 'absolute',
-              top: '45%',
-              left: '1px',
-            },
-          }}
-        >
-          <StyledSideSheetContent>{children}</StyledSideSheetContent>
-        </Resizable>
-      </StyledSideSheet>
-    </StyledScrim>
+    <SideSheetBase {...props}>
+      <StyledContainerWrapper>
+        <StyledHeader>
+          <StyledFlexBox>
+            {components.indicator}
+            <StyledFlexColumn>
+              {components.title}
+              {components.subTitle}
+            </StyledFlexColumn>
+          </StyledFlexBox>
+          <StyledFlexBox>
+            {components.actions}
+            {enableFullscreen && (
+              <Button variant="ghost_icon" onClick={handleFullscreenClick}>
+                <FullscreenIcon />
+              </Button>
+            )}
+            <Button variant="ghost_icon" onClick={onClose}>
+              <Icon name="close" />
+            </Button>
+          </StyledFlexBox>
+        </StyledHeader>
+        {components.content}
+      </StyledContainerWrapper>
+    </SideSheetBase>
   );
 };
+
+SideSheet.Title = Title;
+SideSheet.SubTitle = SubTitle;
+SideSheet.Indicator = Indicator;
+SideSheet.Content = Content;
+SideSheet.Actions = Actions;
 
 export default SideSheet;
