--- packages/side-sheet/src/components/SideSheetCombined.tsx
@@ -1,138 +1,0 @@
-import React, { useCallback } from 'react';
-
-import { Button, Icon } from '@equinor/eds-core-react';
-import { close as closeIcon } from '@equinor/eds-icons';
-import { tokens } from '@equinor/eds-tokens';
-import { type PropsWithChildren, useRef } from 'react';
-import { Actions } from './Actions';
-import { Content } from './Content';
-import { FullscreenIcon } from './icon/FullscreenIcon';
-import { HEXString, Indicator } from './Indicator';
-import { SubTitle } from './SubTitle';
-import { Title } from './Title';
-
-import { SideSheet, type SideSheetProps } from './SideSheet';
-import styled from 'styled-components';
-
-const StyledFlexBox = styled.div`
-  display: flex;
-`;
-
-const StyledFlexColumn = styled(StyledFlexBox)`
-  flex-direction: column;
-`;
-
-const StyledContainerWrapper = styled(StyledFlexColumn)`
-  background: ${tokens.colors.ui.background__default.hex};
-`;
-
-const StyledHeader = styled(StyledFlexBox)`
-  padding: 1rem;
-  height: 48px;
-  justify-content: space-between;
-`;
-
-type PortalSideSheet = SideSheetProps & {
-  minWidth?: number;
-  enableFullscreen?: boolean;
-};
-
-type SideSheetComponents = {
-  indicator?: React.ReactElement<unknown, string | React.JSXElementConstructor<{ color: HEXString }>>;
-  title?: React.ReactElement<unknown, string | React.JSXElementConstructor<{ title: string }>>;
-  subTitle?: React.ReactElement<unknown, string | React.JSXElementConstructor<{ subTitle: string }>>;
-  actions?: React.ReactElement<unknown, string | React.JSXElementConstructor<PropsWithChildren<unknown>>>;
-  content?: React.ReactElement<unknown, string | React.JSXElementConstructor<PropsWithChildren<unknown>>>;
-};
-
-Icon.add({
-  close: closeIcon,
-});
-
-export const SideSheetCombined = (props: PropsWithChildren<PortalSideSheet>) => {
-  const { onClose, children, enableFullscreen } = props;
-
-  const ref = useRef<HTMLDivElement>(null);
-  const handleFullscreenClick = useCallback(() => {
-    if (document.fullscreenElement) {
-      document.exitFullscreen().catch((err) => console.error(err));
-    } else {
-      ref.current?.requestFullscreen();
-    }
-  }, [ref]);
-
-  const components: SideSheetComponents = {
-    indicator: undefined,
-    title: undefined,
-    subTitle: undefined,
-    actions: undefined,
-    content: undefined,
-  };
-
-  React.Children.forEach(children, (child) => {
-    if (!React.isValidElement(child)) return;
-    if (child.type === Indicator) {
-      components.indicator = child;
-    } else if (child.type === Title) {
-      components.title = child;
-    } else if (child.type === SubTitle) {
-      components.subTitle = child;
-    } else if (child.type === Content) {
-      components.content = child;
-    } else if (child.type === Actions) {
-      components.actions = React.cloneElement(child as React.ReactElement<any>, {
-        sideSheetRef: ref,
-      });
-    } else {
-      throw Error(`unsupported child ${child.type} in SideSheet component`);
-    }
-  });
-
-  if (!components.title) {
-    throw Error('Title Component is required child');
-  }
-
-  if (!components.subTitle) {
-    throw Error('SubTitle Component is required child');
-  }
-
-  if (!components.content) {
-    throw Error('Content Component is required child');
-  }
-
-  return (
-    <SideSheet {...props}>
-      <StyledContainerWrapper>
-        <StyledHeader>
-          <StyledFlexBox>
-            {components.indicator}
-            <StyledFlexColumn>
-              {components.title}
-              {components.subTitle}
-            </StyledFlexColumn>
-          </StyledFlexBox>
-          <StyledFlexBox>
-            {components.actions}
-            {enableFullscreen && (
-              <Button variant="ghost_icon" onClick={handleFullscreenClick}>
-                <FullscreenIcon />
-              </Button>
-            )}
-            <Button variant="ghost_icon" onClick={onClose}>
-              <Icon name="close" />
-            </Button>
-          </StyledFlexBox>
-        </StyledHeader>
-        {components.content}
-      </StyledContainerWrapper>
-    </SideSheet>
-  );
-};
-
-SideSheetCombined.Title = Title;
-SideSheetCombined.SubTitle = SubTitle;
-SideSheetCombined.Indicator = Indicator;
-SideSheetCombined.Content = Content;
-SideSheetCombined.Actions = Actions;
-
-export default SideSheetCombined;
