+++ packages/side-sheet/src/index.ts
@@ -1,2 +1,2 @@
-export { SideSheetCombined as SideSheet, default } from './components/SideSheetCombined';
+export { SideSheet, default } from './components/SideSheet';
 export type { HEXString } from './components/Indicator';
